embed_api/requirements.txt
numpy==1.23.3
openai==0.27.8
requests==2.26
uvicorn==0.22.0
scipy==1.10.1
mariadb==1.1.7
pandas==2.0.2
pymongo==4.4.1
tiktoken==0.4.0
PyPDF2==3.0.1
langchain==0.0.149
pydub==0.25.1
fastapi==0.94.0
python-dotenv==0.10.1